Add experimental target 'noqq' (sans inline Quick Quizzes)
Make target "noqq" enables LaTeX boolean "noqqz".
In this case, qqz.sty defines Quick Quiz macros so that inline
Quick Quizzes are typeset as [QQ m.n] (flushed right).

Some of Quick Quizzes are essential to the discussion in the text
and hiding them would make the discussion hard to follow.

As a compromise, annotate such quizzes with a different set of
macros in the form of:

	\EQuickQuiz{
		Quiz part
	}\EQuickQuizAnswer{
		Answer part
	}\EQuickQuizEnd

, and let those quiz parts presented in the text (with background
frames).

As a PoC, annotate Quick Quizzes at the beginning of Chapter "Counting"
using the format above.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>